+++ packages/api-graphql/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 4.0.16 (2024-02-09)
+
+**Note:** Version bump only for package @aws-amplify/api-graphql
+
+## 4.0.15 (2024-02-06)
+
+**Note:** Version bump only for package @aws-amplify/api-graphql
+
 ## 4.0.14 (2024-02-01)
 
 **Note:** Version bump only for package @aws-amplify/api-graphql